requirement.txt
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asteroid-filterbanks==0.4.0
attrs==23.2.0
audioread==3.0.1
av==11.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
ctranslate2==4.1.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.110.1
faster-whisper==1.0.0
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gradio==4.26.0
gradio_client==0.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning==2.2.2
lightning-utilities==0.11.2
llvmlite==0.42.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.17.3
openai==1.17.1
optuna==3.6.1
orjson==3.10.0
packaging==24.0
pandas==2.2.2
pgpt_python==0.1.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
primePy==1.3
protobuf==5.26.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.2
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.7
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.4.0
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
wheel==0.41.2
whisperx @ git+https://github.com/m-bain/whisperx.git@f2da2f858e99e4211fe4f64b5f2938b007827e17
yarl==1.9.4
zipp==3.18.1
